% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results_tables.R
\name{add_baseline_values}
\alias{add_baseline_values}
\title{Helper function}
\usage{
add_baseline_values(base_tbl, baseline_range, step = 100L, seed)
}
\description{
Helper function
}
\keyword{internal}
