% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{add_change_cols}
\alias{add_change_cols}
\title{Add \code{change} and \code{change_pct} columns to a prepared results table}
\usage{
add_change_cols(tbl)
}
\description{
Add \code{change} and \code{change_pct} columns to a prepared results table
}
\keyword{internal}
