% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_principal_stats.R
\name{calculate_principal_stats}
\alias{calculate_principal_stats}
\title{Calculate mean, median, p10 and p90 values from 256 projection model runs}
\usage{
calculate_principal_stats(tbl, group_cols)
}
\arguments{
\item{tbl}{A suitably prepared results data table}

\item{group_cols}{Set of variables to group by, varying slightly for
different inputs/outputs}
}
\value{
A table with baseline values and principal projection summary values
}
\description{
Calculate mean, median, p10 and p90 values from 256 projection model runs
}
\keyword{internal}
