% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results_parquet_files.R
\name{check_grfp_inputs}
\alias{check_grfp_inputs}
\title{Various checks on the input parameters to \code{\link[=get_results_folder_path]{get_results_folder_path()}}}
\usage{
check_grfp_inputs(cn, root_dir, version, scheme, scenario, stamp)
}
\description{
Pulled out to a discrete function, just for neatness' sake
}
\keyword{internal}
