% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results_parquet_files.R
\name{check_single_subdir}
\alias{check_single_subdir}
\title{Aims to return one (1) sub-directory from the given path.}
\usage{
check_single_subdir(path, level, dttm_stamp, container)
}
\arguments{
\item{path}{A single string created by 'file.path': a folder to search within}

\item{level}{Either "scenario" or "dttm_stamp"}

\item{dttm_stamp}{string. If \code{NULL} (the default) and only a single model
run directory is available within the particular combination of \code{version}
\code{scheme} and \code{scenario}, this folder path will be returned. If \code{NULL}, and
multiple model runs are found, an error will be thrown. Alternatively, the
date-time stamp of the desired model run can be supplied. This is expected
to be in the format 'YYYYMMDD_HHMMSS', and must of course match a
subdirectory of the selected 'scenario' folder, else an error will be
thrown. A third alternative is to supply the string \code{"max"}, which will
return the path to the most recent model run, where multiple runs are found}

\item{container}{Azure container. Uses \code{\link[=get_results_container]{get_results_container()}} by default}
}
\value{
A path to an Azure blob storage directory
}
\description{
Conducts various checks and logic to assist the user. Can either return a
scenario directory or a model run (\code{dttm_stamp}) directory
}
\keyword{internal}
