% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_activity_data.R
\name{compile_detailed_activity_data}
\alias{compile_detailed_activity_data}
\title{Prepare data from the \code{sex+age_group} or \code{sex+tretspef_grouped} table}
\usage{
compile_detailed_activity_data(
  results,
  measure,
  tretspef_lookup = get_tretspef_lookup(),
  activity_type = c("ip", "op", "aae"),
  aggregation = c("age_group", "tretspef_grouped"),
  pods = NULL,
  sites = NULL
)
}
\arguments{
\item{results}{A named list containing NHP results tables}

\item{measure}{The measure to focus on for the output table. Valid values
depend on which activity_type is selected}

\item{tretspef_lookup}{A tibble, or a function that returns a tibble,
containing a \code{code} column (used as a key for joining to the tretspef
table) and a \code{tretspef} column that provides friendly labels for specialties}

\item{activity_type}{string. One of "ip", "op", "aae"}

\item{aggregation}{string. One of "age_group" or "tretspef_grouped"}

\item{pods}{character vector. PoD labels to filter data to. The default
value of \code{NULL} means no PoDs will be filtered out}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data}
}
\description{
Prepare data from the \code{sex+age_group} or \code{sex+tretspef_grouped} table
}
