% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{compile_principal_pod_data}
\alias{compile_principal_pod_data}
\title{Prepare data from default results table for displaying as summary table}
\usage{
compile_principal_pod_data(
  results,
  pod_lookup = get_principal_pods(),
  sites = NULL
)
}
\arguments{
\item{results}{A named list containing NHP results tables}

\item{pod_lookup}{A tibble, or a function that returns a tibble, containing
columns named \code{activity_type_label}, \code{pod} and \code{pod_label}. This provides
friendly labels for POD variables in the data. \code{pod} is the key column used
for joining to data tables.}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data}
}
\value{
A filtered and sorted tibble of principal projections of results,
by activity type and point of delivery (PoD)
}
\description{
Prepare data from default results table for displaying as summary table
}
