% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhp_result_sets.R
\name{compile_tagged_runs_metadata_tbl}
\alias{compile_tagged_runs_metadata_tbl}
\title{Generates a table of metadata with a row for each scheme (dataset)}
\usage{
compile_tagged_runs_metadata_tbl(groups = NULL)
}
\arguments{
\item{groups}{Used to pass through any specification to the \code{groups}
argument to \code{get_nhp_user_allowed_datasets()}}
}
\value{
A tibble
}
\description{
The row for each scheme represents the preferred run of any multiple runs,
according to the preference order defined in \code{run_stages()}
}
