% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_activity_type}
\alias{convert_activity_type}
\title{Converts "In|Outpatients", "A&E" strings to "ip", "op" and "aae"}
\usage{
convert_activity_type(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector
}
\description{
Converts "In|Outpatients", "A&E" strings to "ip", "op" and "aae"
}
\keyword{internal}
