% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results_tables.R
\name{create_base_tbl_with_sitetret}
\alias{create_base_tbl_with_sitetret}
\title{Helper function}
\usage{
create_base_tbl_with_sitetret(site)
}
\arguments{
\item{site}{A string giving the name/code of an imaginary site}
}
\description{
Helper function
}
\keyword{internal}
