% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results_tables.R
\name{create_demo_default_tbl}
\alias{create_demo_default_tbl}
\title{Function to create a "default" table with synthetic values}
\usage{
create_demo_default_tbl(seed)
}
\arguments{
\item{seed}{integer, for reproducibility of outputs, a seed value to be used}
}
\description{
The table contains synthetic (randomly created) values for a selection of
PODs and measures, across 2 "sites", as if created by 64 model runs
}
\keyword{internal}
