% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results_tables.R
\name{create_demo_tretspef_losgroup_tbl}
\alias{create_demo_tretspef_losgroup_tbl}
\title{Function to create a "tretspef + LoS group" table with synthetic values
The table contains synthetic (randomly created) values for a selection of
PODs, tretspefs, LoS groups and measures, across 2 "sites", as if created
by 64 model runs}
\usage{
create_demo_tretspef_losgroup_tbl(seed)
}
\arguments{
\item{seed}{integer, for reproducibility of outputs, a seed value to be used}
}
\description{
Function to create a "tretspef + LoS group" table with synthetic values
The table contains synthetic (randomly created) values for a selection of
PODs, tretspefs, LoS groups and measures, across 2 "sites", as if created
by 64 model runs
}
\keyword{internal}
