% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results_tables.R
\name{create_dummy_stepcounts_tbl}
\alias{create_dummy_stepcounts_tbl}
\title{Function to create a "step_counts" table with synthetic values
The table contains synthetic (randomly created) values for a selection of
64 model runs, providing random baselines, random sizes of mitigated
activity across a subset of TPMAs, and random values for demographic growth}
\usage{
create_dummy_stepcounts_tbl(seed)
}
\arguments{
\item{seed}{integer, for reproducibility of outputs, a seed value to be used}
}
\description{
Function to create a "step_counts" table with synthetic values
The table contains synthetic (randomly created) values for a selection of
64 model runs, providing random baselines, random sizes of mitigated
activity across a subset of TPMAs, and random values for demographic growth
}
\keyword{internal}
