% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results_tables.R
\name{create_sample_impacts}
\alias{create_sample_impacts}
\title{Helper function}
\usage{
create_sample_impacts(dat, mitigation, seed, picks = 64)
}
\description{
Helper function
}
\keyword{internal}
