% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exclude_op_teleatt_procedures}
\alias{exclude_op_teleatt_procedures}
\title{Exclude outpatient procedures from tele-attendances count only}
\usage{
exclude_op_teleatt_procedures(tbl)
}
\arguments{
\item{tbl}{A tibble}
}
\description{
Exclude outpatient procedures from tele-attendances count only
}
\keyword{internal}
