% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_activity_data.R
\name{export_detailed_activity_data}
\alias{export_detailed_activity_data}
\title{Prepare a site-level summary of detailed activity results by PoD and LoS}
\usage{
export_detailed_activity_data(
  results,
  tretspef_lookup = get_tretspef_lookup(),
  aggregation = c("age_group", "tretspef"),
  sites = NULL
)
}
\arguments{
\item{results}{A named list containing NHP results tables}

\item{tretspef_lookup}{A tibble, or a function that returns a tibble,
containing a \code{code} column (used as a key for joining to the tretspef
table) and a \code{tretspef} column that provides friendly labels for specialties}

\item{aggregation}{string. One of "age_group" or "tretspef"}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data}
}
\value{
A tibble
}
\description{
Intended to be used to create a table to be exported to .csv/.xlsx
}
