% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_factor_data.R
\name{export_principal_cf_data}
\alias{export_principal_cf_data}
\title{Prepare a site-level summary table of change_factor results}
\usage{
export_principal_cf_data(
  results,
  sites = NULL,
  pod_lookup = get_detailed_pods(),
  tpma_lookup = get_tpma_label_lookup()
)
}
\arguments{
\item{results}{A named list containing NHP results tables}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data}

\item{pod_lookup}{A tibble, or a function that returns a tibble, containing
columns named \code{activity_type_label}, \code{pod} and \code{pod_label}. This provides
friendly labels for POD variables in the data. \code{pod} is the key column used
for joining to data tables.}

\item{tpma_lookup}{A tibble, or a function that returns a tibble, containing
a column named \code{strategy} (used as a key for joining to the \code{step_counts}
table) and a column named \code{tpma_label} that provides friendly labels for
all TPMAs/strategies}
}
\value{
A tibble
}
\description{
Intended to be used to create a table to be exported to .csv/.xlsx
}
