% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_los_data.R
\name{export_site_level_principal_los_data}
\alias{export_site_level_principal_los_data}
\title{Prepare a site-level summary of main projection results by PoD and LoS}
\usage{
export_site_level_principal_los_data(los_tbl, sites = NULL)
}
\arguments{
\item{los_tbl}{the "tretspef+los_group" table from NHP results}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data.}
}
\value{
A tibble
}
\description{
Intended to be used to create a table to be exported to .csv/.xlsx
}
