% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{format_bar_cols}
\alias{format_bar_cols}
\title{Format horizontal \code{gt_bar}s within tables}
\usage{
format_bar_cols(tbl, p_col = "principal", p_clr = "#686f73")
}
\description{
Format horizontal \code{gt_bar}s within tables
}
\keyword{internal}
