% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{format_gt_core}
\alias{format_gt_core}
\title{Common helper function to handle standar formatting of gt tables}
\usage{
format_gt_core(gt_table, extra_col = NULL)
}
\description{
Common helper function to handle standar formatting of gt tables
}
\keyword{internal}
