% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_results_tables.R
\name{generate_horizon_values}
\alias{generate_horizon_values}
\title{Helper function}
\usage{
generate_horizon_values(baseline, mitigation, seed, picks = 64)
}
\description{
Helper function
}
\keyword{internal}
