% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_activity_type_from_pod}
\alias{get_activity_type_from_pod}
\title{Create a column called \code{activity_type} by extracting a substring from \code{pod}}
\usage{
get_activity_type_from_pod(tbl)
}
\arguments{
\item{tbl}{A tibble}
}
\description{
Create a column called \code{activity_type} by extracting a substring from \code{pod}
}
\keyword{internal}
