% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_github_app_installation_id}
\alias{get_github_app_installation_id}
\title{Returns the installation ID for the reskit GitHub app}
\usage{
get_github_app_installation_id(jwt = get_github_jwt())
}
\arguments{
\item{jwt}{JWT for the app. Defaults to the output of \code{get_github_jwt()}.}
}
\value{
An installation ID as an integer
}
\description{
The envvars "RESKIT_GH_APP_ID" and "RESKIT_GH_APP_PRIVATE_KEY" must
be set. The former as a string, the latter as a path to a PEM key.
}
\keyword{internal}
