% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_github_jwt}
\alias{get_github_jwt}
\title{Returns a GitHub JWT for the reskit authentication app}
\usage{
get_github_jwt()
}
\value{
A JSON Web Token (JWT) as a string
}
\description{
The envvars "RESKIT_GH_APP_ID" and "RESKIT_GH_APP_PRIVATE_KEY" must
be set. The former as a string, the latter as a path to a PEM key.
}
\keyword{internal}
