% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_github_pat}
\alias{get_github_pat}
\title{Return a GitHub access token}
\usage{
get_github_pat()
}
\value{
A personal access token (PAT) with permissions granted to the app
}
\description{
If the envvar "RESKIT_GH_APP_PAT" is set, this will be used as the token.
Otherwise, the envvars "RESKIT_GH_APP_ID" and "RESKIT_GH_APP_PRIVATE_KEY"
must be set. The former as a string, the latter as a path to a PEM key.
}
\keyword{internal}
