% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_github_pat_via_api}
\alias{get_github_pat_via_api}
\title{Return a GitHub PAT via the API using the reskit authentication GitHub app}
\usage{
get_github_pat_via_api()
}
\value{
An access token (PAT) as a string
}
\description{
The envvars "RESKIT_GH_APP_ID" and "RESKIT_GH_APP_PRIVATE_KEY" must
be set. The former as a string, the latter as a path to a PEM key.
}
\keyword{internal}
