% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhp_user_allowed_datasets.R
\name{get_nhp_user_allowed_datasets}
\alias{get_nhp_user_allowed_datasets}
\title{Return a vector of results datasets that the user is allowed to access}
\usage{
get_nhp_user_allowed_datasets(groups = NULL)
}
\arguments{
\item{groups}{A vector of groups that the user belongs to. \code{NULL} by default.}
}
\value{
A character vector
}
\description{
Return a vector of results datasets that the user is allowed to access
}
