% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_outputs_gh_file}
\alias{get_outputs_gh_file}
\title{Read in a file from the NHP Outputs app GitHub repo}
\usage{
get_outputs_gh_file(...)
}
\arguments{
\item{...}{Pass the name of the file in via \code{...}}
}
\description{
Read in a file from the NHP Outputs app GitHub repo
}
\keyword{internal}
