% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_principal_pods.R
\name{get_pod_measures_lst}
\alias{get_pod_measures_lst}
\title{Read in list of PoD labelling data from local yaml file}
\usage{
get_pod_measures_lst()
}
\description{
Read in list of PoD labelling data from local yaml file
}
\keyword{internal}
