% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhp_user_allowed_datasets.R
\name{get_providers}
\alias{get_providers}
\title{Return a vector of provider codes from the supporting data container}
\usage{
get_providers()
}
\value{
A character vector
}
\description{
Return a vector of provider codes from the supporting data container
}
