% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{get_results_container}
\alias{get_results_container}
\title{Get an Azure storage container for results data}
\usage{
get_results_container(rcv_name = "AZ_RESULTS_CONTAINER")
}
\arguments{
\item{rcv_name}{The name of the \emph{environment variable} that stores the
name of the results container. "AZ_RESULTS_CONTAINER" by default.}
}
\value{
An Azure storage container
}
\description{
Get an Azure storage container for results data
}
\examples{
\dontrun{  get_results_container() }
}
