% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results_parquet_files.R
\name{get_results_dir_path}
\alias{get_results_dir_path}
\title{Return a path to a specific results directory}
\usage{
get_results_dir_path(
  container,
  path,
  version,
  scheme,
  scenario = NULL,
  dttm_stamp = NULL
)
}
\arguments{
\item{container}{An Azure container.}

\item{path}{string. The path to the directory within the container, where the
desired results data is stored}

\item{version}{string. The NHP model version in the format "v3.6", or "dev"}

\item{scheme}{string. The code/name of the NHP scheme. May be "national" etc}

\item{scenario}{string. If \code{NULL}, the default, and only a single scenario is
available within the particular combination of \code{version} and \code{scheme}, this
single path will be followed. However if multiple scenarios are available,
an error will be thrown. The user may avert this by simply supplying an
existent scenario name}

\item{dttm_stamp}{string. If \code{NULL} (the default) and only a single model
run directory is available within the particular combination of \code{version}
\code{scheme} and \code{scenario}, this path will be returned. If \code{NULL}, and
multiple model runs are found, an error will be thrown. Alternatively, the
date-time stamp of the desired model run can be supplied. This is expected
to be in the format 'YYYYMMDD_HHMMSS', and must of course match a
subdirectory of the selected 'scenario' directory, else an error will be
thrown. A third alternative is to supply the string \code{"max"}, which will
return the path to the most recent model run, where multiple runs are found}
}
\value{
A path to an Azure blob storage directory, as a string
}
\description{
Provides some logic and checks to assist the user in locating a valid
directory from which to read in NHP results data
}
