% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results_parquet_files.R
\name{get_results_folder_path}
\alias{get_results_folder_path}
\title{Specify a results folder path}
\usage{
get_results_folder_path(
  version,
  scheme,
  scenario = NULL,
  dttm_stamp = NULL,
  container = NULL,
  root_dir = NULL
)
}
\arguments{
\item{version}{string. The NHP model version in the format "v3.6", or "dev"}

\item{scheme}{string. The code/name of the NHP scheme. May be "national" etc}

\item{scenario}{string. If \code{NULL}, the default, and only a single scenario is
available within the particular combination of \code{version} and \code{scheme}, this
single folder path will be followed. However if multiple scenarios are
available, an error will be thrown. The user may avert this by simply
supplying an existent scenario name}

\item{dttm_stamp}{string. If \code{NULL} (the default) and only a single model
run directory is available within the particular combination of \code{version}
\code{scheme} and \code{scenario}, this folder path will be returned. If \code{NULL}, and
multiple model runs are found, an error will be thrown. Alternatively, the
date-time stamp of the desired model run can be supplied. This is expected
to be in the format 'YYYYMMDD_HHMMSS', and must of course match a
subdirectory of the selected 'scenario' folder, else an error will be
thrown. A third alternative is to supply the string \code{"max"}, which will
return the path to the most recent model run, where multiple runs are found}

\item{container}{Azure container. Uses \code{\link[=get_results_container]{get_results_container()}} by default}

\item{root_dir}{string. The name of the root folder within the container
where the all desired results data is stored. Uses the value of the
environment variable "AZ_RESULTS_DIRECTORY" by default}
}
\value{
A path to an Azure blob storage directory, as a string
}
\description{
Provides some logic and checks to assist the user in locating
a valid folder from which to read in NHP results data
}
