% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_su_gh_file}
\alias{get_su_gh_file}
\title{Read in a file from a Strategy Unit GitHub repo}
\usage{
get_su_gh_file(repo, folder, file, pat = get_github_pat())
}
\arguments{
\item{repo}{string. The name of the repository in which to find the file}

\item{folder}{string. The folder where the file is located. Set to \code{""} to
use the root folder of the repo.}

\item{file}{string. The name of the file to read in}

\item{pat}{A GitHub Personal Access Token. Generated by \link{get_github_pat} by
default, via a custom Strategy Unit app, but you might pass in a personal
token if you prefer or if that pipeline is not working. If you have a token
stored as an environment variable called "RESKIT_GH_APP_PAT", this will be
be used, saving the need to make any API calls.}
}
\value{
<<<<<<< HEAD
The URL to the raw file contents, to be passed to a reader function
=======
The file contents as a stream, to be passed to a reader function
>>>>>>> 7d456db (:otter: Refactor functions that pull lookup files from GitHub)
}
\description{
Read in a file from a Strategy Unit GitHub repo
}
\keyword{internal}
