% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/containers.R
\name{get_support_container}
\alias{get_support_container}
\title{Get an Azure storage container for supporting data}
\usage{
get_support_container(scv_name = "AZ_SUPPORT_CONTAINER")
}
\arguments{
\item{scv_name}{The name of the \emph{environment variable} that stores the
name of the supporting data container. "AZ_SUPPORT_CONTAINER" by default.}
}
\value{
An Azure storage container
}
\description{
Get an Azure storage container for supporting data
}
\examples{
\dontrun{  get_support_container() }
}
