% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_tretspef_lookup}
\alias{get_tretspef_lookup}
\title{Get a lookup of tretspef codes to descriptions}
\usage{
get_tretspef_lookup()
}
\value{
A 2-column tibble with columns \code{code} and \code{tretspef}
}
\description{
Get a lookup of tretspef codes to descriptions
}
