% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_trust_sites}
\alias{get_trust_sites}
\title{From any results table, get list of all site codes for this scheme}
\usage{
get_trust_sites(res_tbl, col = "sitetret")
}
\arguments{
\item{res_tbl}{A tibble from the results list}

\item{col}{string The name of the column containing site codes. \code{sitetret} by
default}
}
\value{
A character vector
}
\description{
The "default" table is recommended
}
