% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhp_result_sets.R, R/reskit-package.R
\name{gregg}
\alias{gregg}
\title{grepl a glued regex}
\usage{
gregg(x, rx, g = parent.frame())

gregg(x, rx, g = parent.frame())
}
\value{
A logical value
}
\description{
Facilitates using regex in search/filter patterns, and puts the
arguments "the right way round" (x first, then pattern), unlike \code{grepl()}

Facilitates using regex in search/filter patterns, and puts the
arguments "the right way round" (x first, then pattern), unlike \code{\link[=grepl]{grepl()}}
}
\keyword{internal}
