% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhp_result_sets.R
\name{gregv}
\alias{gregv}
\title{grepv a glued regex}
\usage{
gregv(x, rx, g = parent.frame())
}
\value{
All values of x that match the regex in rx
}
\description{
Facilitates using regex in search/filter patterns, and puts the
arguments "the right way round" (x first, then pattern), unlike \code{\link[=grepv]{grepv()}}
}
\keyword{internal}
