% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{gt_bar}
\alias{gt_bar}
\title{Function to handle the size and formatting of gt_bar elements in tables}
\usage{
gt_bar(x, format_fn = NULL, colours = c("#ec6555", "#f9bf07"))
}
\description{
Function to handle the size and formatting of gt_bar elements in tables
}
\keyword{internal}
