% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{inner_join_for_labels}
\alias{inner_join_for_labels}
\title{Use a lookup table to get more readable labels for PoDs}
\usage{
inner_join_for_labels(tbl, lookup)
}
\arguments{
\item{tbl}{A tibble}

\item{lookup}{A lookup table with pod and pod_label columns}
}
\description{
Use a lookup table to get more readable labels for PoDs
}
\keyword{internal}
