% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{join_for_labels}
\alias{join_for_labels}
\title{Use a lookup table to get more readable labels for PoDs}
\usage{
join_for_labels(tbl, lookup)
}
\arguments{
\item{tbl}{A tibble}

\item{lookup}{A lookup table with \code{pod} and \code{pod_label} columns}
}
\description{
Use a lookup table to get more readable labels for PoDs
}
\keyword{internal}
