% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{keep_measures}
\alias{keep_measures}
\title{A vector of values from the measure column in results tables}
\usage{
keep_measures()
}
\description{
Currently this contains 6 of the 7 possible values; it excludes "procedures".
This vector is used in several places as a filter.
}
\keyword{internal}
