% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plots.R
\name{make_cumulative_distrib_plot}
\alias{make_cumulative_distrib_plot}
\title{Generate a cumulative distribution function (ecdf) chart of model run values}
\usage{
make_cumulative_distrib_plot(distrib_plot_data, show_zero = FALSE)
}
\arguments{
\item{distrib_plot_data}{data frame. As produced by
\link{compile_distribution_plot_data}}

\item{show_zero}{Boolean. Whether to extend the chart to show a zero value,
for context.}
}
\description{
Generate a cumulative distribution function (ecdf) chart of model run values
}
