% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_plots.R
\name{make_grouped_impact_plot}
\alias{make_grouped_impact_plot}
\title{Generate overall change factor ("waterfall") chart}
\usage{
make_grouped_impact_plot(grouped_impact_data)
}
\arguments{
\item{grouped_impact_data}{tibble, as produced by
\link{compile_grouped_impact_data}}
}
\description{
Generate overall change factor ("waterfall") chart
}
