% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_principal_pod_table}
\alias{make_principal_pod_table}
\title{Create a gt table with data from \link{compile_principal_pod_data}}
\usage{
make_principal_pod_table(principal_pod_data)
}
\arguments{
\item{principal_pod_data}{A suitable tibble}
}
\value{
A gt table
}
\description{
Create a gt table with data from \link{compile_principal_pod_data}
}
