% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_plots.R
\name{make_tpma_impact_plot}
\alias{make_tpma_impact_plot}
\title{Generate bar charts by change factor at individual TPMA level}
\usage{
make_tpma_impact_plot(tpma_impact_data)
}
\arguments{
\item{tpma_impact_data}{tibble, as produced by \link{compile_tpma_impact_data}}
}
\description{
Generate bar charts by change factor at individual TPMA level
}
