% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_activity_data.R
\name{prepare_age_group_data}
\alias{prepare_age_group_data}
\title{Prepare data from the 'sex+age_group' results table}
\usage{
prepare_age_group_data(results)
}
\arguments{
\item{results}{List of NHP results tables}
}
\description{
Prepare data from the 'sex+age_group' results table
}
\keyword{internal}
