% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_los_data.R
\name{prepare_site_level_principal_los_data}
\alias{prepare_site_level_principal_los_data}
\title{Initial preparation of data for the main LoS summary table}
\usage{
prepare_site_level_principal_los_data(los_tbl, selected_measure)
}
\arguments{
\item{los_tbl}{the "tretspef+los_group" table from NHP results}
}
\value{
A tibble
}
\description{
Initial preparation of data for the main LoS summary table
}
\keyword{internal}
