% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{prepare_site_level_principal_pod_data}
\alias{prepare_site_level_principal_pod_data}
\title{Initial preparation of data for the main summary table}
\usage{
prepare_site_level_principal_pod_data(default_tbl)
}
\arguments{
\item{default_tbl}{the "default" table from NHP results}
}
\value{
A tibble
}
\description{
Initial preparation of data for the main summary table
}
\keyword{internal}
