% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_activity_data.R
\name{prepare_tretspef_data}
\alias{prepare_tretspef_data}
\title{Prepare data from the 'sex+tretspef_grouped' results table}
\usage{
prepare_tretspef_data(results, tretspef_lookup)
}
\arguments{
\item{results}{A named list containing NHP results tables}

\item{tretspef_lookup}{A tibble, or a function that returns a tibble,
containing a \code{code} column (used as a key for joining to the tretspef
table) and a \code{tretspef} column that provides friendly labels for specialties}
}
\description{
Prepare data from the 'sex+tretspef_grouped' results table
}
\keyword{internal}
