% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_activity_data.R
\name{proto_detailed_activity_data}
\alias{proto_detailed_activity_data}
\title{Zero-row prototype for the \link{compile_detailed_activity_data} output}
\usage{
proto_detailed_activity_data(aggregation)
}
\arguments{
\item{aggregation}{character Either "tretspef" or "age_group"}
}
\value{
A zero-row tibble
}
\description{
The column names and types here must match what
\link{compile_detailed_activity_data} returns when rows are present
}
\keyword{internal}
