% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_data.R
\name{proto_export_impact_data}
\alias{proto_export_impact_data}
\title{Zero-row prototype for the \link{export_impact_data} function output}
\usage{
proto_export_impact_data()
}
\value{
A zero-row tibble
}
\description{
Zero-row prototype for the \link{export_impact_data} function output
}
\keyword{internal}
