% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_data.R
\name{proto_grouped_impact_data}
\alias{proto_grouped_impact_data}
\title{Zero-row prototype for the \link{compile_grouped_impact_data} function output}
\usage{
proto_grouped_impact_data()
}
\value{
A zero-row tibble
}
\description{
The column names and types here must match what
\link{compile_grouped_impact_data} returns when rows are present
\code{test-empty_results.R} asserts this.
}
\keyword{internal}
