% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhp_result_sets.R
\name{read_all_tagged_runs_params}
\alias{read_all_tagged_runs_params}
\title{Read in data from the \code{params.json} file for each tagged runS}
\usage{
read_all_tagged_runs_params(runs_metadata_tbl = NULL)
}
\arguments{
\item{runs_metadata_tbl}{Tibble of metadata, as produced by
\link{compile_tagged_runs_metadata_tbl}. If not supplied,
compile_tagged_runs_metadata_tbl() will be run anyway to create the table.}
}
\value{
A list of run params data, each named with its scheme code
}
\description{
Read in data from the \code{params.json} file for each tagged runS
}
