% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results_parquet_files.R
\name{read_results_parquet_files}
\alias{read_results_parquet_files}
\title{Read a selection of (or all) parquet files in an Azure directory}
\usage{
read_results_parquet_files(container, path, tables = NULL)
}
\arguments{
\item{container}{An Azure container.}

\item{path}{string. Path to an Azure directory of results data. Potentially
produced by \link{get_results_dir_path}}

\item{tables}{character vector. \code{NULL}, the default, results in all available
parquet files in the \code{path} folder being read in. If you wish only to read
in a subset of the files, specify their names here, without the ".parquet"
file extension}
}
\value{
A named list of tibbles
}
\description{
Read a selection of (or all) parquet files in an Azure directory
}
\examples{
\dontrun{
  data <- azkit::get_container("data_container") |>
    read_results_parquet_files("data/dev/national/test", "acuity")

  data_list <- azkit::get_container("data_container") |>
    get_results_dir_path("files", version = "v4.0", scheme = "national") |>
    read_results_parquet_files()
}
}
