% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{read_tpmas_lookup}
\alias{read_tpmas_lookup}
\title{Read the TPMAs lookup table from a CSV file in the TPMAs repo}
\usage{
read_tpmas_lookup(file = "tpma-lookup.csv")
}
\arguments{
\item{file}{The name of the csv file to be read in}
}
\description{
Read the TPMAs lookup table from a CSV file in the TPMAs repo
}
\keyword{internal}
