% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{read_tx_lookup}
\alias{read_tx_lookup}
\title{Read the tretspef lookup table from a JSON file in the NHP Outputs repo}
\usage{
read_tx_lookup(path = file.path("app", "data", "tx-lookup.json"))
}
\arguments{
\item{path}{The path (relative to \verb{inst/}) to the json file to be read in}
}
\description{
Read the tretspef lookup table from a JSON file in the NHP Outputs repo
}
\keyword{internal}
