% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{relabel_pods}
\alias{relabel_pods}
\alias{relabel_ip_activity_types}
\title{Give PoDs more accurate labels}
\usage{
relabel_pods(tbl)

relabel_ip_activity_types(tbl)
}
\arguments{
\item{tbl}{A tibble}
}
\description{
Give PoDs more accurate labels

Give activity types more accurate labels
}
\keyword{internal}
