% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reskit-package.R
\docType{package}
\name{reskit-package}
\alias{reskit}
\alias{reskit-package}
\title{reskit: Nice selection box of goodies for hobnobbing with NHP model results}
\description{
'Wafer' goodbye to crumbly code, and dunk a sturdy reskit function in your data cuppa instead. We've clustered the cream of our results code here - it's a cracker!
}
\author{
\strong{Maintainer}: Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})

}
\keyword{internal}
