% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shim_results.R
\name{shim_results}
\alias{shim_results}
\title{Wrangle tables from old-style results to match format of new-style results}
\usage{
shim_results(results)
}
\arguments{
\item{results}{List of old style NHP results tables}
}
\value{
List of NHP results tables in the "new" format
}
\description{
Wrangle tables from old-style results to match format of new-style results
}
