% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plot_data.R
\name{compile_distribution_plot_data}
\alias{compile_distribution_plot_data}
\title{Compile data to support the "activity distribution summary" tables}
\usage{
compile_distribution_plot_data(
  default_tbl,
  measure,
  activity_type = c("ip", "op", "aae"),
  pods = NULL,
  sites = NULL
)
}
\arguments{
\item{measure}{The measure to focus on for the output table. Valid values
depend on which activity_type is selected}

\item{activity_type}{string. One of "ip", "op", "aae"}

\item{pods}{character vector. PoD labels to filter data to. The default
value of \code{NULL} means no PoDs will be filtered out}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data}
}
\value{
A tibble
}
\description{
Compile data to support the "activity distribution summary" tables
}
