% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_summary_data.R
\name{compile_distribution_summary_data}
\alias{compile_distribution_summary_data}
\title{Compile data to support the "activity distribution summary" tables}
\usage{
compile_distribution_summary_data(
  default_tbl,
  value_type = c("median", "principal"),
  sites = NULL
)
}
\arguments{
\item{default_tbl}{the "default" table from NHP results}

\item{value_type}{string Either "median" or "principal"}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data}
}
\value{
A tibble
}
\description{
Compile data to support the "activity distribution summary" tables
}
