% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{compile_principal_pod_data}
\alias{compile_principal_pod_data}
\title{Prepare data from default results table for displaying as summary table}
\usage{
compile_principal_pod_data(default_tbl, sites = NULL)
}
\arguments{
\item{default_tbl}{the "default" table from NHP results}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data.}
}
\value{
A filtered and sorted tibble of principal projections of results,
by activity type and point of delivery (PoD)
}
\description{
Prepare data from default results table for displaying as summary table
}
