% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhp_result_sets.R
\name{compile_run_metadata_tbl}
\alias{compile_run_metadata_tbl}
\title{Compiles a table of metadata parameters for each model run for each scheme}
\usage{
compile_run_metadata_tbl(groups = NULL)
}
\arguments{
\item{groups}{Used to pass through any specification to the \code{groups}
argument to \code{get_nhp_user_allowed_datasets()}}
}
\value{
A tibble
}
\description{
This function uses data from the "aggregated-model-results" Azure folder.
}
