% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_summary_data.R
\name{export_distribution_summary_data}
\alias{export_distribution_summary_data}
\title{Prepare a site-level summary of activity distribution summary data}
\usage{
export_distribution_summary_data(default_tbl)
}
\arguments{
\item{default_tbl}{the "default" table from NHP results}
}
\value{
A tibble
}
\description{
Intended to be used to create a table to be exported to .csv/.xlsx
}
