% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_factor_data.R
\name{export_principal_cf_data}
\alias{export_principal_cf_data}
\title{Prepare a site-level summary table of change_factor results}
\usage{
export_principal_cf_data(stepcounts_tbl, sites = NULL)
}
\arguments{
\item{stepcounts_tbl}{The "step_counts" table from NHP results}

\item{sites}{Either \code{NULL} (the default) or a vector of site codes to filter
to. \code{NULL} means don't filter; include all sites present in the data}
}
\value{
A tibble
}
\description{
Intended to be used to create a table to be exported to .csv/.xlsx
}
