% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_to_main_measures}
\alias{filter_to_main_measures}
\title{Filter a table so the \code{measure} column only contains 6 selected measures}
\usage{
filter_to_main_measures(tbl)
}
\arguments{
\item{tbl}{A tibble}
}
\description{
Currently this contains 6 of the 7 possible values; it excludes "procedures".
This function is used in several places in reskit as a filter.
}
\keyword{internal}
