% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_principal_pods.R
\name{get_detailed_pods}
\alias{get_detailed_pods}
\alias{get_principal_pods}
\title{Prepare a lookup table with activity type labels and PoD labels for each PoD}
\usage{
get_detailed_pods()

get_principal_pods()
}
\value{
A tibble

A tibble
}
\description{
This function condenses all A&E activity to a single category - compare
\code{get_detailed_pods()} which keeps all A&E categories
}
