% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_principal_pods.R
\name{get_principal_pods}
\alias{get_principal_pods}
\title{Prepare a lookup table with activity type labels and PoD labels for each PoD}
\usage{
get_principal_pods()
}
\value{
A tibble
}
\description{
Prepare a lookup table with activity type labels and PoD labels for each PoD
}
\keyword{internal}
