% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_tpma_label_lookup}
\alias{get_tpma_label_lookup}
\title{Get a lookup of TPMA labels}
\usage{
get_tpma_label_lookup()
}
\value{
A 2-column tibble, with columns \code{strategy} and \code{tpma_label}
}
\description{
Currently reads from a fixed location within the package.
}
