% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_tpma_label_lookup}
\alias{get_tpma_label_lookup}
\title{Get a lookup of TPMA labels from a 'strategy' variable}
\usage{
get_tpma_label_lookup(file = here::here("inst/mitigators.json"))
}
\arguments{
\item{file}{Path to JSON source file}
}
\value{
A 2-column tibble
}
\description{
Get a lookup of TPMA labels from a 'strategy' variable
}
