% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_file.R
\name{get_tpmas_gh_file}
\alias{get_tpmas_gh_file}
\title{Read in a file from the TPMAs GitHub repo}
\usage{
get_tpmas_gh_file(file, folder = "reference")
}
\arguments{
\item{file}{string. The name of the file to read in}

\item{folder}{string. The folder where the file is located. \code{"reference"} by
default. Set to \code{""} to use the root folder of the repo.}
}
\value{
The file contents as a stream, to be passed to a reader function
}
\description{
Read in a file from the TPMAs GitHub repo
}
\keyword{internal}
