% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{inner_join_for_labels}
\alias{inner_join_for_labels}
\title{Use a lookup table to get more readable labels for PoDs}
\usage{
inner_join_for_labels(tbl, lookup)
}
\description{
Use a lookup table to get more readable labels for PoDs
}
\keyword{internal}
