% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_detailed_activity_table}
\alias{make_detailed_activity_table}
\title{Create a gt table with data from compile_detailed_activity_data()}
\usage{
make_detailed_activity_table(detailed_activity_data, final_year)
}
\arguments{
\item{detailed_activity_data}{A suitable tibble}

\item{final_year}{string. The horizon year eg "2042/43"}
}
\value{
A gt table
}
\description{
Create a gt table with data from compile_detailed_activity_data()
}
