% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_distribution_summary_table}
\alias{make_distribution_summary_table}
\title{Create a gt table with data from compile_distribution_summary_data()}
\usage{
make_distribution_summary_table(distr_summary_data)
}
\arguments{
\item{distr_summary_data}{A suitable tibble}
}
\value{
A gt table
}
\description{
Create a gt table with data from compile_distribution_summary_data()
}
