% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_factor_plots.R
\name{make_individual_cf_plot}
\alias{make_individual_cf_plot}
\title{Generate bar charts by change factor at individual TPMA level}
\usage{
make_individual_cf_plot(indiv_change_factor_data, x_axis_label)
}
\arguments{
\item{indiv_change_factor_data}{data frame. As produced by
\link{compile_indiv_change_factor_data}}

\item{x_axis_label}{string. A label for the charts' x-axis.}
}
\description{
Generate bar charts by change factor at individual TPMA level
}
