% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_factor_plots.R
\name{make_overall_cf_plot}
\alias{make_overall_cf_plot}
\title{Generate overall change factor ("waterfall") chart}
\usage{
make_overall_cf_plot(principal_change_factor_data)
}
\arguments{
\item{principal_change_factor_data}{data frame. As produced by
\link{compile_change_factor_data}}
}
\description{
Generate overall change factor ("waterfall") chart
}
