% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_activity_data.R
\name{prepare_detailed_activity_data}
\alias{prepare_detailed_activity_data}
\title{Data preparation step for 'activity in detail' table}
\usage{
prepare_detailed_activity_data(init_data, aggregation)
}
\value{
A tibble
}
\description{
Data preparation step for 'activity in detail' table
}
\keyword{internal}
