% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_activity_data.R
\name{prepare_detailed_activity_data}
\alias{prepare_detailed_activity_data}
\title{Data preparation step for 'activity in detail' table}
\usage{
prepare_detailed_activity_data(dat, aggregation)
}
\arguments{
\item{aggregation}{string. One of "age_group" or "tretspef_grouped"}
}
\value{
A tibble
}
\description{
Data preparation step for 'activity in detail' table
}
\keyword{internal}
