% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_plot_data.R
\name{prepare_distribution_plot_data}
\alias{prepare_distribution_plot_data}
\title{Preparation of site-level data for the main summary table}
\usage{
prepare_distribution_plot_data(default_tbl)
}
\arguments{
\item{default_tbl}{the "default" table from NHP results}
}
\value{
A tibble
}
\description{
Preparation of site-level data for the main summary table
}
\keyword{internal}
