% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_los_data.R
\name{prepare_principal_los_data}
\alias{prepare_principal_los_data}
\title{Preparation of site-level data for the main LoS summary table}
\usage{
prepare_principal_los_data(dat)
}
\arguments{
\item{dat}{A tibble}
}
\value{
A tibble
}
\description{
Preparation of site-level data for the main LoS summary table
}
\keyword{internal}
