% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results_parquet_files.R
\name{read_results_parquet_files}
\alias{read_results_parquet_files}
\title{Read a selection of (or all) parquet files in an Azure directory}
\usage{
read_results_parquet_files(path, tables = NULL, container = NULL)
}
\arguments{
\item{path}{string. Path to an Azure directory of results data. As produced
by \code{\link[=get_results_folder_path]{get_results_folder_path()}}}

\item{tables}{character vector. \code{NULL}, the default, results in all available
parquet files in the \code{path} folder being read in. If you wish only to read
in a subset of the files, specify these here, without the ".parquet" file
extension}

\item{container}{Azure container. Uses \code{\link[=get_results_container]{get_results_container()}} by default}
}
\value{
A named list of tibbles
}
\description{
Read a selection of (or all) parquet files in an Azure directory
}
\examples{
\dontrun{
  data <- read_results_parquet_files("data/dev/national/test", "acuity")

  get_results_folder_path(version = "v4.0", scheme = "national") |>
    read_results_parquet_files()
}
}
