% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_pod_data.R
\name{relabel_pods}
\alias{relabel_pods}
\title{Convert PoDs and activity types to more accurate labels}
\usage{
relabel_pods(tbl)
}
\description{
Convert PoDs and activity types to more accurate labels
}
\keyword{internal}
